Update initial variable introduction to be "name"
At slide 37, students are introduced to the variable `name` and are encouraged to throughout slide 37 to 41 to discover how variables are used using the `fullname` variable.

Students face issues with the follow on exercise at slide 42, when the variable that they were using now change to `name` and cause their `document.write(name)` statement to not work.

This commit change all reference to `firstname` variable to be `name` for simplified course follow on and introducing the `firstname` variable not in camel case notation confuses students when the next chapter introduces camel case notation. 

A line break is also introduced to slide 41 to separate both statements under point #2 due to students writing `var firstname = ? var firstname = prompt("What is your name?");` in one line.